improved test & pre-compilation
+ test/test.py:
   - add path setup so that one could run test.py without setting pysolid as python module (via path setup or pip)
   - save figs in test dir
   - open figs in macOS and linux for visual checking

+ test/precompile_solid.sh:
   - exit the shell if any command fail
   - run test.py after building
   - print passing msg

+ solid.cpy*.so: update witth numppy version constraint for improved compatibility
   - add test/conda*.txt for pre-compilation with numpy version constraint of a upper limit, to maximize the compatibility, because the solid.cpython*.so file generated from lower numpy version works for envs with higher numpy versions.
   - update solid.cpython*.so files for macOS/Linux

+ grid: more robust call of numpy.loadtxt()

+++ pysolid/grid.py
@@ -97,7 +97,7 @@ def calc_solid_earth_tides_grid(dt_obj, atr, step_size=1e3, display=False, verbo
                     usecols=(2,3,4),
                     delimiter=',',
                     skiprows=0,
-                    max_rows=length*width)
+                    max_rows=int(length*width))
     tide_e = fc[:, 0].reshape(length, width)
     tide_n = fc[:, 1].reshape(length, width)
     tide_u = fc[:, 2].reshape(length, width)
